Apply Scaling Factor to VIIRS Data

Apply scaling factor to variables according to Black Marble user guide. The scaling factor is 0.1 for specific VIIRS variables.

Usage

apply_scaling_factor_to_viirs_data(x, variable, quiet = TRUE)

Arguments

x

A numeric vector or matrix representing the VIIRS data.

variable

A character string specifying the variable name.

quiet

Logical, if TRUE, suppresses warnings when the variable is not found in scaling variables.

Value

A numeric vector or matrix with the scaling factor applied to the specified variables.

Details

This function applies a scaling factor of 0.1 to specific VIIRS variables according to the Black Marble user guide.

The following VIIRS variables are affected:

DNB_At_Sensor_Radiance (VNP46A1)

DNB_BRDF-Corrected_NTL (VNP46A2)

Gap_Filled_DNB_BRDF-Corrected_NTL (VNP46A2)

DNB_Lunar_Irradiance (VNP46A2)

AllAngle_Composite_Snow_Covered (VNP46A3/4)

AllAngle_Composite_Snow_Covered_Std (VNP46A3/4)

AllAngle_Composite_Snow_Free (VNP46A3/4)

AllAngle_Composite_Snow_Free_Std (VNP46A3/4)

NearNadir_Composite_Snow_Covered (VNP46A3/4)

NearNadir_Composite_Snow_Covered_Std (VNP46A3/4)

NearNadir_Composite_Snow_Free (VNP46A3/4)

NearNadir_Composite_Snow_Free_Std (VNP46A3/4)

OffNadir_Composite_Snow_Covered (VNP46A3/4)

OffNadir_Composite_Snow_Covered_Std (VNP46A3/4)

OffNadir_Composite_Snow_Free (VNP46A3/4)

OffNadir_Composite_Snow_Free_Std (VNP46A3/4)

References

Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf