Black Marble Tiles Spatial Data

Spatial dataset containing black marble tiles obtained from NASA's Black Marble project.

Usage

black_marble_tiles_sf

Format

A spatial dataframe (sf) with 648 rows and 2 columns:

TileID

Identification of the tile associated with the VIIRS satellites mapping.

geometry

Spatial geometry representing the tile boundaries.

Source

https://blackmarble.gsfc.nasa.gov/

The dataset is obtained from the Black Marble project by NASA, specifically the VIIRS (Visible Infrared Imaging Radiometer Suite) satellites. The TileID column represents the identification of the tile associated with the VIIRS satellites mapping. For more details on the dataset and its characteristics, refer to the Black Marble User Guide: https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf

References

Tan, Bin, et al. "Night-time light imagery analysis for urban population mapping in China." International Journal of Remote Sensing 27.9 (2006): 1943-1950.

Wolfe, Robert E., et al. "Achieving accuracy requirements for forest biomass mapping: A data fusion method for combining the strengths of spaceborne lidar and Landsat TM data." Remote Sensing of Environment 80.2 (2002): 253-267.

See also

Examples

head(black_marble_tiles_sf)
#> Simple feature collection with 6 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: 30.00012 xmax: -170 ymax: 90.00012
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 2
#>   TileID                                                                geometry
#>   <chr>                                                            <POLYGON [°]>
#> 1 h00v00 ((-180 90.00012, -170 90.00012, -170 80.00012, -180 80.00012, -180 90.…
#> 2 h00v01 ((-180 80.00012, -170 80.00012, -170 70.00012, -180 70.00012, -180 80.…
#> 3 h00v02 ((-180 70.00012, -170 70.00012, -170 60.00012, -180 60.00012, -180 70.…
#> 4 h00v03 ((-180 60.00012, -170 60.00012, -170 50.00012, -180 50.00012, -180 60.…
#> 5 h00v04 ((-180 50.00012, -170 50.00012, -170 40.00012, -180 40.00012, -180 50.…
#> 6 h00v05 ((-180 40.00012, -170 40.00012, -170 30.00012, -180 30.00012, -180 40.…