Extracts raster data for specified region of interest

Extracts raster data for specified region of interest

Usage

bm_extract(
  roi_sf,
  product_id,
  date,
  bearer,
  aggregation_fun = c("mean"),
  add_n_pixels = TRUE,
  variable = NULL,
  quality_flags_to_remove = NULL,
  check_all_tiles_exist = TRUE,
  interpol_na = FALSE,
  output_location_type = "memory",
  file_dir = NULL,
  file_prefix = NULL,
  file_skip_if_exists = TRUE,
  quiet = FALSE,
  ...
)

Arguments

roi_sf

Spatial object representing the region of interest.

product_id

Identifier for the product.

date

Dates for which raster data is to be extracted.

bearer

Authentication token for accessing the data.

aggregation_fun

Aggregation function to summarize raster values (default is "mean").

add_n_pixels

Logical indicating whether to add pixel count information (default is TRUE).

variable

Variable to extract from the product (default is NULL).

quality_flags_to_remove

Quality flag removal options (default is NULL).

check_all_tiles_exist

Logical indicating whether to check if all tiles exist (default is TRUE).

interpol_na

Logical indicating whether to interpolate missing values (default is FALSE).

output_location_type

Output location type ("memory" or "file", default is "memory").

file_dir

Directory for saving output files (used if output_location_type is "file", default is NULL).

file_prefix

Prefix for output file names (default is NULL).

file_skip_if_exists

Logical indicating whether to skip file creation if output file already exists (default is TRUE).

quiet

Logical indicating whether to suppress progress messages (default is FALSE).

Value

A data frame containing extracted raster data.