Make Black Marble Raster

Make a raster of nighttime lights from NASA Black Marble data

bm_raster(roi_sf, product_id, date, bearer, variable = NULL)

Arguments

roi_sf

Region of interest; sf polygon. Must be in the WGS 84 (epsg:4326) coordinate reference system.

product_id

One of the following:

"VNP46A1": Daily (raw)

"VNP46A2": Daily (corrected)

"VNP46A3": Monthly

"VNP46A4": Annual

date

Date of raster data. Entering one date will produce a raster. Entering multiple dates will produce a raster stack.

For product_ids "VNP46A1" and "VNP46A2", a date (eg, "2021-10-03").

For product_id "VNP46A3", a date or year-month (e.g., "2021-10-01", where the day will be ignored, or "2021-10").

For product_id "VNP46A4", year or date (e.g., "2021-10-01", where the month and day will be ignored, or 2021).

bearer

NASA bearer token. For instructions on how to create a token, see here.

variable

Variable to used to create raster (default: NULL). If NULL, uses the following default variables:

For product_id :VNP46A1", uses DNB_At_Sensor_Radiance_500m.

For product_id "VNP46A2", uses Gap_Filled_DNB_BRDF-Corrected_NTL.

For product_ids "VNP46A3" and "VNP46A4", uses NearNadir_Composite_Snow_Free. For information on other variable choices, see here; for VNP46A1, see Table 3; for VNP46A2 see Table 6; for VNP46A3 and VNP46A4, see Table 9.

Value

Raster

Examples

if (FALSE) {
# Define bearer token
bearer <- "BEARER-TOKEN-HERE"

# sf polygon of Ghana
library(geodata)
roi_sf <- gadm(country = "GHA", level=0, path = tempdir()) %>% st_as_sf()

# Daily data: raster for October 3, 2021
ken_20210205_r <- bm_raster(roi_sf = roi_sf,
                            product_id = "VNP46A2",
                            date = "2021-10-03",
                            bearer = bearer)

# Monthly data: raster for March 2021
ken_202103_r <- bm_raster(roi_sf = roi_sf,
                          product_id = "VNP46A3",
                          date = "2021-03-01",
                          bearer = bearer)

# Annual data: raster for 2021
ken_2021_r <- bm_raster(roi_sf = roi_sf,
                        product_id = "VNP46A4",
                        date = 2021,
                        bearer = bearer)
}