Download VIIRS Satellite Image in HDF5 Format

Downloads a VIIRS satellite image in HDF5 format from NASA's LADSWeb and saves it to a temporary directory or persistent location.

Usage

download_h5_viirs_sat_image(
  file_name,
  temp_dir,
  bearer,
  quality_flags_to_remove = numeric(),
  quiet = FALSE
)

Arguments

file_name

A character string representing the name of the file to download.

temp_dir

A character string specifying the temporary directory where the file will be saved.

bearer

A character string containing the authorization token for accessing NASA's LADSWeb.

quality_flags_to_remove

A numeric vector containing quality flag values to be removed from the data (optional).

quiet

Logical; indicating whether to suppress progress messages (default: FALSE).

Details

This function downloads a VIIRS satellite image in HDF5 format from NASA's LADSWeb based on the provided file name. It then processes the downloaded data, removing quality flag values if specified. The function also provides an option to suppress progress messages.

References

Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf